Server Config helper
The middleware 'setup' was removed. All the logic for server configuration will be placed in this file.

Morgan is setup wo work with winston logger. We remove a new line from morgan message in order to avoid an extra new line in the final log message.

Also is configured static content: a public folder for a client app and an API documentation file.