Account component
All CRUD operations for the Accounts API are addressed here. It requeries a property containing the loged in Firebase Auth token in order to make requests to the API.

The pagination is handled by v-data-table component from vuetify.

This components get all the accounts items when is mounted and show API error messages in the add/edit dialog.