New folder structure
Due to some issues with now.sh and herouko services with the previous folder structure I decided to change the folder structure. In this new structure the root folder contains all the service files. Client project is added as a sub project of Service. NPM scripts were modified to match the new folder structure.